requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.12.0
django-cors-headers>=3.7.0
djangorestframework-simplejwt>=5.0.0